Refactor InspectionStart component
Refactor InspectionStart.tsx into smaller, more manageable files for better code organization and maintainability.
